+++ crates/solidity/src/solc/soljson_compiler.rs
@@ -1,4 +1,4 @@
-//! The Solidity compiler.
+//! The Solidity compiler solJson interface.
 
 use std::path::Path;
 use std::path::PathBuf;
@@ -18,22 +18,30 @@ extern "C" {
 }
 
 /// The Solidity compiler.
-pub struct SoljsonCompiler {
-    /// The lazily-initialized compiler version.
-    pub version: Option<Version>,
-}
+pub struct SoljsonCompiler;
 
 impl Compiler for SoljsonCompiler {
     /// Compiles the Solidity `--standard-json` input into Yul IR.
     fn standard_json(
         &mut self,
         mut input: StandardJsonInput,
-        _base_path: Option<String>,
-        _include_paths: Vec<String>,
-        _allow_paths: Option<String>,
+        base_path: Option<String>,
+        include_paths: Vec<String>,
+        allow_paths: Option<String>,
     ) -> anyhow::Result<StandardJsonOutput> {
-        let version = self.version()?;
-        input.normalize(&version.default);
+        if !include_paths.is_empty() {
+            anyhow::bail!("configuring include paths is not supported with solJson")
+        }
+        if base_path.is_some() {
+            anyhow::bail!("configuring the base path is not supported with solJson")
+        }
+        if allow_paths.is_some() {
+            anyhow::bail!("configuring allow paths is not supported with solJson")
+        }
+
+        let version = self.version()?.validate(&include_paths)?.default;
+        input.normalize(&version);
+
         let suppressed_warnings = input.suppressed_warnings.take().unwrap_or_default();
 
         let input_json = serde_json::to_string(&input).expect("Always valid");
@@ -74,31 +82,11 @@ impl Compiler for SoljsonCompiler {
             .ok_or_else(|| anyhow::anyhow!("Soljson version parsing: metadata dropping"))?
             .parse()
             .map_err(|error| anyhow::anyhow!("Soljson version parsing: {}", error))?;
-
         let l2_revision: Option<semver::Version> = version
             .split('-')
             .nth(1)
             .and_then(|version| version.parse().ok());
-
-        let version = Version::new(long, default, l2_revision);
-        if version.default < super::FIRST_SUPPORTED_VERSION {
-            anyhow::bail!(
-                "`Soljson` versions <{} are not supported, found {}",
-                super::FIRST_SUPPORTED_VERSION,
-                version.default
-            );
-        }
-        if version.default > super::LAST_SUPPORTED_VERSION {
-            anyhow::bail!(
-                "`Soljson` versions >{} are not supported, found {}",
-                super::LAST_SUPPORTED_VERSION,
-                version.default
-            );
-        }
-
-        self.version = Some(version.clone());
-
-        Ok(version)
+        Ok(Version::new(long, default, l2_revision))
     }
 }